[skip ci] Add template for Sanity pipelines (#28981)
This PR does not affect APC or any existing pipeline.

This PR is the first of a series of refactoring PRs for pipeline
reorganization. Please refer to
https://tenstorrent.atlassian.net/wiki/spaces/MI6/pages/1396506680/Proposed+pipeline+and+test+organization+changes
for more details.

Todo list beyond this PR is currently tracked at
https://github.com/tenstorrent/tt-metal/issues/30675

Goals of this refactoring:

- create simplified interface for devs adding tests 
- clearly communicate machine time allocation via time_budget.yaml
- auto verify time budgets at the beginning of workflows
- clearer ownership of tests
- less pipelines to get a PR merged

Using the `tests.yaml` file to adjust timeouts:

- each test must specify timeout and sku (currently only N300 and P150b
supported for specifically ops tests)
- the sum of the timeout per sku must be less than or equal to the
specified budget in `time_budget.yaml`

Testing:

- [x] APC calling Ops Sanity
https://github.com/tenstorrent/tt-metal/actions/runs/18585877776
(failure on one job expected)
- [x] Ops Sanity workflow dispatch
https://github.com/tenstorrent/tt-metal/actions/runs/18585828695